% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmap.R
\name{cnmap}
\alias{cnmap}
\title{Mapping China use amCharts + R}
\usage{
cnmap(id = NULL, value = NULL, description = NULL, theme = NULL,
  colorSteps = 12, smallMap = 1, smallMapTop = 50,
  smallMapRight = 50, export = 1, valueLegendRight = 10,
  valueLegendBottom = 10, minValue = "min", maxValue = "max",
  title = "Map of China", titleSize = 18, titleColor = "#363636",
  subtitle = "czxa.top", subtitleSize = 16,
  subtitleColor = "#363636", autoZoom = 1, outlineAlpha = 0.5,
  outlineColor = "#363636", outlineThickness = 0.5,
  rollOverColor = NULL, ..., width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{id}{province name}

\item{value}{province value}

\item{description}{detailed description}

\item{theme}{theme: black/chalk/dark/light}

\item{colorSteps}{Number of layers of colors in legend}

\item{smallMap}{Weather to display a minimap or not. smallMap = 1 means to display, smallMao = 0 means to not display.}

\item{smallMapTop}{Control the vertical position of the minimap.}

\item{smallMapRight}{Control the Horizontal position of the minimap.}

\item{export}{Whether to open the export plug-in, 1 means to open, 0 means not to open.}

\item{valueLegendRight}{Horizontal position of legend}

\item{valueLegendBottom}{Vertical position of legend}

\item{minValue}{Label of Minimum Legend}

\item{maxValue}{Label of Maximum Legend}

\item{title}{title}

\item{titleSize}{title size}

\item{titleColor}{title color}

\item{subtitle}{subtitle}

\item{subtitleSize}{subtitle size}

\item{subtitleColor}{subtitle color}

\item{autoZoom}{Whether to zoom in automatically or not, 1 means yes and 0 means no.}

\item{outlineAlpha}{Transparency of borders}

\item{outlineColor}{Color of borders}

\item{outlineThickness}{Width of borders}

\item{rollOverColor}{The color of the area where the mouse passes}

\item{...}{NULL}

\item{width}{NULL}

\item{height}{NULL}

\item{elementId}{NULL}
}
\description{
Mapping China use amCharts + R
}
\examples{
library(dplyr)
library(magrittr)
library(hchinamap)
library(mapbrew)
df <- chinadf \%>\% dplyr::filter(region == "中国")
cnmap(id = df$name, value = df$value,
     description = paste0(df$name, ": ", df$value, "人"),
     smallMap = 0, theme = "dark")

}
